Refactor cache and model to use integer for mtime, enhance SQLiteMetadataCache with set_entries method, and update FSWalker to use mtime in nanoseconds

+++ src/fspachinko/domain/model.py
@@ -110,7 +110,7 @@ class FSEntry:
     ext: str
     parent: str
     size: int
-    mtime: float
+    mtime: int
     duration: float = 0.0
